width

+++ sketch/matrix.h
@@ -20,8 +20,7 @@ const uint8_t* generate_matrix(const float& x, const float& y) {
   static uint8_t led[WIDTH * HEIGHT];
   memset(led, 0, sizeof(led));
 
-  int xx =
-      int((std::min(std::max(y, -1.0f), 1.0f) + 1.0f) / 2.0f * (WIDTH - 1));
+  int xx = (std::min(std::max(y, -1.0f), 1.0f) + 1.0f) / 2.0f * (WIDTH - 1);
   int yy = int(std::min(std::max(x, 0.0f), 1.0f) * (HEIGHT - 1));
 
   led[xx] = 1;
